+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added an icon to link to embedded asset URLs when an embedded asset is selected in an assets field (inspired by the Embedded Assets Link plugin for Craft 3)
+- Added the `showFieldLinkIcon` plugin setting (defaults to `true`) to control whether or not to show field link icons
 
 ## 3.1.9 - 2023-11-24